requirements.txt
flask==3.0.0
requests==2.31.0
openai==1.12.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
lxml==5.1.0
